Add Lock/Unlock App
Add Lock/Unlock App with biometric or screenlock credentials
